- replaced double used when calculating Fibonacci with BigInteger. Double was used to prevent from integer overflow, but whole numbers have a representation with floating point in Web.config
- create the backapp project
- initial commit containing the project
- added FibonacciController with Get and Post(not implemented)
- web api config to return html not xml response
- added unit tests from consoleapp
- added CustomMath.cs from consoleapp

+++ consoleapp/FibonacciConsoleApp/FibonacciConsoleApp/Program.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Numerics;
 using CustomMath;
 
 namespace FibonacciMultiplicationTableConsoleApp
@@ -23,7 +24,7 @@ namespace FibonacciMultiplicationTableConsoleApp
                     }
                     else
                     {
-                        List<List<double>> multiplicationTable = Fibonacci.GetMultiplicationTable(size);
+                        List<List<BigInteger>> multiplicationTable = Fibonacci.GetMultiplicationTable(size);
                         Console.WriteLine("  " + String.Join(" ", multiplicationTable[0].ToArray()));
                         for (int i = 1; i < multiplicationTable.Count; i++)
                         {
